avoid overflow

+++ src/dict.c
@@ -408,7 +408,7 @@ static dictEntry *dictGenericDelete(dict *d, const void *key, int nofree) {
 
     for (table = 0; table <= 1; table++) {
         idx = h & DICTHT_SIZE_MASK(d->ht_size_exp[table]);
-        if (table == 0 && idx < d->rehashidx) continue;
+        if (table == 0 && (long)idx < d->rehashidx) continue;
         he = d->ht_table[table][idx];
         prevHe = NULL;
         while(he) {
@@ -517,7 +517,7 @@ dictEntry *dictFind(dict *d, const void *key)
     h = dictHashKey(d, key);
     for (table = 0; table <= 1; table++) {
         idx = h & DICTHT_SIZE_MASK(d->ht_size_exp[table]);
-        if (table == 0 && idx < d->rehashidx) continue; 
+        if (table == 0 && (long)idx < d->rehashidx) continue; 
         he = d->ht_table[table][idx];
         while(he) {
             if (key==he->key || dictCompareKeys(d, key, he->key))
@@ -1090,7 +1090,7 @@ dictEntry **dictFindEntryRefByPtrAndHash(dict *d, const void *oldptr, uint64_t h
     if (dictSize(d) == 0) return NULL; /* dict is empty */
     for (table = 0; table <= 1; table++) {
         idx = hash & DICTHT_SIZE_MASK(d->ht_size_exp[table]);
-        if (table == 0 && idx < d->rehashidx) continue;
+        if (table == 0 && (long)idx < d->rehashidx) continue;
         heref = &d->ht_table[table][idx];
         he = *heref;
         while(he) {
